+++ backend/src/main/java/com/ssafy/api/request/dto/Live/LiveCategoryDto.java
@@ -10,13 +10,13 @@ import javax.persistence.Id;
 public class LiveCategoryDto {
     @Id
     @Column(name="live_pk")
-    Long livepk;
+    Integer livepk;
     @Column(name="product_pk")
-    Long productpk;
+    Integer productpk;
     @Column(name="live_title")
     String livetitle;
     @Column(name="live_viewer_count")
-    Long liveviewercount;
+    Integer liveviewercount;
     @Column(name="user_id")
     String userid;
     String filepath;
