+++ backend/src/main/java/com/ssafy/db/entity/Image.java
@@ -13,7 +13,7 @@ public class Image{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long imageId;
+    private Integer imageId;
 
     @ManyToOne(fetch = FetchType.LAZY,cascade = CascadeType.ALL)
     @JoinColumn(name = "product_pk")
