+++ backend/src/main/java/com/ssafy/db/repository/Live/LiveRepositorySupport.java
@@ -27,7 +27,7 @@ public class LiveRepositorySupport {
     QImage qImage = QImage.image;
     QProduct qProduct = QProduct.product;
 
-    public Tuple findByLiveId(Long liveid){
+    public Tuple findByLiveId(int liveid){
         Tuple live =  jpaQueryFactory.select(qLive,qUser)
                 .from(qLive)
                 .join(qUser).on(qLive.userid.eq(qUser.userid)).where(qLive.livepk.eq(liveid))
@@ -42,12 +42,12 @@ public class LiveRepositorySupport {
         return live;
     }
 
-    public Long endLive(Long value){
+    public Long endLive(int value){
         //Live테이블에 islive0값으로 update
-        Long a = jpaQueryFactory.update(qLive).set(qLive.islive,0L)
+        Long a = jpaQueryFactory.update(qLive).set(qLive.islive,0)
                 .where(qLive.livepk.eq(value)).execute();
         //Product 테이블에 isLive 0값으로 update
-        Long b = jpaQueryFactory.update(qProduct).set(qProduct.isLive,0L)
+        Long b = jpaQueryFactory.update(qProduct).set(qProduct.isLive,0)
                 .where(qProduct.liveId.eq(value)).execute();
         return a;
     }
@@ -60,7 +60,7 @@ public class LiveRepositorySupport {
 
     public List<LiveMainDto> selectMain(){
         List<Live> l = jpaQueryFactory.select(qLive).from(qLive)
-                .where(qLive.islive.eq(Long.valueOf(1)))
+                .where(qLive.islive.eq(1))
                 .orderBy(qLive.liveviewercount.desc()).limit(12).fetch();
         List<LiveMainDto> res = new LinkedList<>();
         for(Live i : l){
@@ -71,7 +71,7 @@ public class LiveRepositorySupport {
         return res;
     }
 
-    public List<LiveCategoryDto> getLiveByCategoryId(Long categoryid){
+    public List<LiveCategoryDto> getLiveByCategoryId(int categoryid){
         List<Tuple> l = jpaQueryFactory.select(qLive,qProduct).from(qLive)
                 .join(qProduct).on(qProduct.id.eq(qLive.productpk))
                 .where(qProduct.isSold.eq(0),qProduct.categoryId.eq(categoryid)).fetch();
